refactor : LottoList class 명 LottoTicket으로 변경 및 LottoList List Type으로 변경
- List가 실제 리스트가 아니라면 List라는 단어를 명명하지 않는다.
- Array에서 형변환 및 데이터 관리가 편한 List 형으로 변경
